requirements.txt
fastapi[standard]
pydantic
networkx
matplotlib
python-dotenv
